Make socket handle parameter uint64_t
Add error messages to list
Rework some of the logic

+++ src/tppSocketPOSIX.h
@@ -8,7 +8,7 @@ namespace tpp
 
 		SocketPOSIX();
 
-		SocketPOSIX(int socketHandle);
+		SocketPOSIX(uint64_t socketHandle);
 
 		virtual ISocket* Accept(const NetworkAddress& address) override;
